Add resiliance to BggApi calls
Handle problems with getting board game info in a basic fashion rather than not at all.  These changes will prevent us from having a valid game entry if a valid bgg game id is supplied but we can't talk to BGG for some reason, that will need to be fixed.